Remove explicit dependencies of Pearl.
[UPDATED]   No longer using UIColor+Expanded in Pearl-UIKit.
[UPDATED]   No longer requiring JRSwizzle to compile Pearl.
[UPDATED]   The base Pearl and UIKit module now no longer have any required dependencies.

+++ Pearl/PearlCodeUtils.m
@@ -103,7 +103,7 @@ uint64_t PearlSecureRandom() {
 
         memset(decodingTable, CHAR_MAX, 256);
         for (char i = 0; i < 64; i++)
-            decodingTable[CodeUtils_Base64EncodingTable[i]] = i;
+            decodingTable[(int)CodeUtils_Base64EncodingTable[(int)i]] = i;
     }
 
     const char *characters = [self cStringUsingEncoding:NSASCIIStringEncoding];
